Add working server commands
Currently, count is bugged due to incorrect assumptions on repeat names
with simultaneous tf_bot_add.  In the meantime, count should remain at 1
until a clean naming convention that will not cause conflicts is arrived.
Therefore, all classes in warbands.txt should have unique names.  If
these rules are not observed, the bot autokick on death will not be valid.